satvi_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
graphene
graphene_sqlalchemy
